pyproject.toml
[tool.poetry]
name = "intensive_django"
version = "0.1.0"
description = ""
authors = ["h4x4d <sidenkoleg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
django = ">=3.2.0,<3.3.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
dlint = "^0.13.0"

[tool.poetry.group.test]
optional = false


[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pyhamcrest = "^2.0.4"
allure-pytest = "^2.11.1"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
